refactor(entities-plugins): register custom_key through FieldRenderer
`ExpressionField` is the field component for an expressible field, so a plugin
customizing one registers it like any other component — the shape
`openid-connect` uses for its own fields:

    <FieldRenderer v-slot="slotProps" :match="({ path }) => path === 'config.custom_key'">
      <CustomKeyField v-bind="slotProps" />
    </FieldRenderer>

`CustomKeyField` is `ExpressionField` plus the wording the counter key needs,
since its expression overrides the key rather than a limit. It lives with
rate-limiting-advanced and rate-limiting imports it, rather than being copied
into both.

Answers the review question on whether `fieldRenderers` work for an expressible
field: `ExpressionField.spec.ts` registers it for one path with
`propsOverrides` and asserts the field renders with a single editor while its
siblings keep the shared one.

No change to `Field.vue` or the dispatch: a registered renderer replaces the
field, and here the renderer is that field's own component.
