+++ packages/entities/entities-plugins/src/components/free-form/plugins/rate-limiting/RateLimitingForm.vue
@@ -1,10 +1,18 @@
 <template>
   <DynamicLayout
     v-bind="props"
-    :field-renderers="fieldRenderers"
     :render-rules="renderRules"
     :schema="gatedSchema"
   >
+    <template #field-renderers>
+      <FieldRenderer
+        v-slot="slotProps"
+        :match="({ path }) => path === 'config.custom_key'"
+      >
+        <CustomKeyField v-bind="slotProps" />
+      </FieldRenderer>
+    </template>
+
     <ConfigForm />
   </DynamicLayout>
 </template>
@@ -13,12 +21,13 @@
 import { AUTOFILL_SLOT, AUTOFILL_SLOT_NAME } from '@kong-ui-public/forms'
 import { provide } from 'vue'
 import ConfigForm from './ConfigForm.vue'
-import CustomKeyField from '../_shared/CustomKeyField.vue'
+import CustomKeyField from '../rate-limiting-advanced/CustomKeyField.vue'
+import FieldRenderer from '../../shared/FieldRenderer.vue'
 import DynamicLayout from '../../shared/layout/DynamicLayout.vue'
 import { useExpressionMode } from '../_shared/use-expression-mode'
 
 import type { PluginFormLayoutProps as Props } from '../../shared/layout/provider'
-import type { FieldRenderer, RenderRules } from '../../shared/types'
+import type { RenderRules } from '../../shared/types'
 
 const props = defineProps<Props>()
 
@@ -27,18 +36,6 @@ const props = defineProps<Props>()
 // `second`..`year` and `custom_key` are expressible; gate their expression
 // editors with the rest of the 3.16 features, leaving the fields themselves.
 const { gatedSchema } = useExpressionMode(() => props.schema)
-
-/**
- * `custom_key` overrides the counter key rather than a limit, so its expression
- * gets its own example and help text. Registered rather than placed, so the
- * field keeps its position among the auto-rendered siblings.
- */
-const fieldRenderers: FieldRenderer[] = [
-  {
-    match: 'config.custom_key',
-    component: CustomKeyField,
-  },
-]
 
 // A custom component owns its own render rules, so these live here rather than
 // in the plugin config.
